refactor: ♻️ extract lifecycle methods to lifecycle.rs (#191)
